Add HRMET code to repository including README, instructions, etc.
One minor bug fix occurred since the published version of the code,
related to the calculation of clf (sky cloud fraction). In the
version of the code used in Zipper & Loheide (2014), the cloud fraction
was inadvertently being set to 0 in most cases where it should have been
between 0-1. Since HRMET input data was primarily collected under
clear-sky conditions, this does not have a big impact on the results.